HTTP/2 stream id should be unsigned int
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpcore/trunk@1739459 13f79535-47bb-0310-9956-ffa450edef68

+++ httpcore5-h2/src/main/java/org/apache/hc/core5/http2/frame/ByteBufferFrame.java
@@ -32,7 +32,7 @@ public final class ByteBufferFrame extends Frame<ByteBuffer> {
 
     private final ByteBuffer payload;
 
-    public ByteBufferFrame(final int type, final int flags, final long streamId, final ByteBuffer payload) {
+    public ByteBufferFrame(final int type, final int flags, final int streamId, final ByteBuffer payload) {
         super(type, flags, streamId);
         this.payload = payload;
     }
